[terminal] IJPL-215236 Refactor: rename terminal command completion action
Better to not mention `Gen2` in the action ID because it will stick with us forever. While gen2 is a temporary term.
And also update its user-visible name by adding a verb 'Call'. Now it should be similar to the editor one 'Call Inline Completion'.
Add text override for Search Everywhere.

GitOrigin-RevId: d32e5b18e8e8e1ddd7e3cf7d8f013a174eed934c