[python] PY-84902 Fix uv run configuration for RemDev
For some reason the way we run Remote Development from Toolbox
initializes some different environment with super default PATH variable
without reading ~/.profile. That's why uv was not found in PATH and we
weren't able to run anything with it.

It's safer to detect uv executable properly in known paths rather than
just passing "uv" and rely on proper environment.

GitOrigin-RevId: f9abfda97977dc29999f5d374f96cc213be7a8a2