[platform UI] IJPL-189308 Fix tool window backgrounds in remdev
The main change here is the introduction of
UIUtil.getGeneralizedParentOfType.
It takes foster parents into account, thus taking care
of exotic Lux cases that were the reason why
it didn't work for some tool windows in remdev.
Instead of relying on the islands holder,
the more generic InternalDecorator is used here,
which is either the usual InternalDecoratorImpl
or its remdev counterpart BackendToolwindowHeaderDecorator.
Thus, this logic works in both monolith and remdev.

Also, the checks for SHOWING_CHANGED
and component.isShowing were added,
which reduces the overhead, as otherwise the listener
is triggered too often, on every parent change,
even before the component is shown.

Also to reduce the overhead, the forEachComponentInHierarchy
call is removed because a separate event is fired for every
component anyway.


(cherry picked from commit 3137f441d5dca95aad58d2d1a9ac3801d37bf355)

IJ-CR-180119

GitOrigin-RevId: e11c72e9a60378c34b362b41f6f035009fd49fdf