[go] GO-19392 add an assertion to tests for the Go parser that validates the structural identity between the light AST and PSI
This check is important to be sure that the Go parser produces same AST structure (`ASTNode`) in PSI as in case of light AST structure (`LigherASTNode`). It defends us against the theoretical inconsistency when a lazy element is mishandled: for example, the light tree stores flat state, but PSI tree parses deeper. Some other cases also possible.

GitOrigin-RevId: c9afab6ec5ed9358d7739b1dd4a5684b0fb2e423