[eel][ijent] IJPL-212328 EDT FS checker complains on actual I/O and allows instantiation of Path (reworked)
Previously, `EelEarlyAccessChecker` produced a lot of false-positive warnings because it reacted on instantiation of `java.nio.file.Path`. However, it is a harmless wrapper around strings.

Now the error message is written only when something actually tries to initialize IJent in EDT.

It is the first attempt to fix the problem. THe first commit 5e9e089083dc223674cad8f3a0692706b96afacf had so many bugs that it was easier to revert it and rewrite from zero.

GitOrigin-RevId: e0ee077d5787e4904342e69b187a521c764cb41d