[air] lsp: dynamic capability registration
This patchset adds dynamic capability registration.

Dynamic capabilities are implemented in a form of dynamic features, by
conforming to the `DynamicFeature` interface (and its corresponding
`Descriptor`).

Symmetrically, one can create "static" features by conforming to the
`StaticFeature` interface. This is mostly mimicking the vscode model.

As an example, two features were implemented:
`LSPPublishDiagnosticFeature` which is a "static" feature, and
`LSPDiagnosticFeature`, which is an example of a "dynamic" feature.

Dynamic features can be "activated" by registering them. The
registration can happen upon server initialisation, if the server
responds to the initialised request with registration data in its
capabilities. If not, the server may later invoke the
`"registerCapability"` method to register a dynamic feature.

Dynamic features can be later "deactivated" if the server invokes the
`"unregisterCapability"` method with the corresponding registration id.

This is the starting point for converting more features to allow their
dynamic registration, which is a requirement for some servers.

GitOrigin-RevId: 5ad44219ed08546ab200d7b210d8176c5d45bf8b