[UI] IJPL-189623 Introduce keys for stripe button text offsets
This determines the absolute shift of the text rectangle.
Therefore, the local variable textOffset is renamed to textPadding
as it wasn't exactly an offset, to avoid confusion.

The default offsets are zero because texts are normally centered.
But for island themes we'll need this shift to compensate
for the island's border.

GitOrigin-RevId: 4cd028116d262d8ad2edd0f8c4cda6e4c2332329