[UI] IJPL-189623 Introduce keys for stripe button text offsets
This determines the absolute shift of the text rectangle.
Therefore, the local variable textOffset is renamed to textPadding
as it wasn't exactly an offset, to avoid confusion.

The default offsets are zero because texts are normally centered.
But for island themes we'll need this shift to compensate
for the island's border.


(cherry picked from commit 4cd028116d262d8ad2edd0f8c4cda6e4c2332329)

IJ-CR-179392

GitOrigin-RevId: 885136fa0cdb292756f19e59ae1b07ab69e3cb5a