[UI] IJPL-192439 Islands: background image support
Enable background image painting by disabling the gradient
for now.

To make it work, we stop setting the NO_BACKGROUND
property if the background image is enabled.
For now this requires a restart or islands on/off toggle.
We'll have to fix this later.

The editor with the empty text requires a bit of special
handling because it can use different backgrounds
for the empty text itself and for the surrounding island.

Another issue is that for the time being we disable
the gradient globally if at least one open project
has the background image.

GitOrigin-RevId: 69322c3136f5c1c01680b229d7953f1f2db8f2a6