[UI] IJPL-215925 Islands: reduce tool window stripe width
To make it aligned with the settings button on macOS,
instead of centering the icons in the visually wider stripes,
we reduce the visual stripe width and shift everything to the edge.
This creates an illusion of the stripe having the same width
as with non-island themes.

The Compact Mode is unaffected because by chance
everything is already aligned there.

Update the UI test to reflect the new width.


(cherry picked from commit 8d107d2dfa3acf5f78ea09816273a5e6df4df8fe)

IJ-CR-180803

GitOrigin-RevId: b4cd695210ef4eae448bfc1c4c47e16d256e0ed6