[wsl] don't run wsl.exe in the frontend process to collect statistics about WSL usage (IJPL-214732)
WSL may be restricted on systems which are running JetBrains Client, so it's better not to call it just to collect statistics.

It would be better to move the collector to a module which isn't loaded in the frontend process instead, but for now the minimal fix is implemented to simplify backporting.

GitOrigin-RevId: 52ed9f43b26b3e18ee7f260ed5c70db969e7fcd1