[wsl] don't run wsl.exe in the frontend process to collect statistics about WSL usage (IJPL-214732)
WSL may be restricted on systems which are running JetBrains Client, so it's better not to call it just to collect statistics.

It would be better to move the collector to a module which isn't loaded in the frontend process instead, but for now the minimal fix is implemented to simplify backporting.


(cherry picked from commit bdd1387ca259adbef5b5e7d93747eaeeaab93339)

IJ-MR-180845

GitOrigin-RevId: c0d6fb3213678e0235e37c098c025f9a9ab3e9a8