[UI, Wayland] IJPL-156745 Dispose popups when they're hidden
When Wayland decides to hide the window because its
coordinates became "illegal," it actually hides it,
not closes. This makes the whole platform consider
the popup as still active. Let's close and dispose it then.
This fixes, among other things, the inability to invoke SE
after it was hidden by Wayland.

GitOrigin-RevId: 17839cd29ab5365f77bdd8556d29652933bca134