IJPL-215965 IJent: failure in obtaining env var doesn't ruin the whole session anymore
The problem was in this incorrect code:

```
 fun environmentVariables(opts: EelExecApi.EnvironmentVariablesOptions): Deferred<Map<String, String>> =
    environmentVariablesCache
      .updateAndGet { cached ->
        if (...) {
          cached
        }
        else {
          coroutineScope.async(start = CoroutineStart.LAZY) {
```

Here, `coroutineScope` is the scope assigned for the whole session with the IJent. It is intentionally not a supervisor scope. Any error thrown inside this coroutine terminates the IJent process.

Execution of ~/.bashrc can trigger an error. Examples: a long-running command that doesn't fit the 30-second timeout, or an explicit call of `exit`.


(cherry picked from commit 2bc3f727b765cba505bc21a08b44e2bcbf32574d)

IJ-CR-181035

GitOrigin-RevId: c42f0eae97d6b86c65bed5e401f7236a31b527cf