[gradle][kotlin][tests] IDEA-302835 simplify test fixture for version catalogs: catalogs could be empty
Before, to support navigation from build.gradl.kts to version catalog elements, it was required to sync version catalogs with their content. Because navigation relied on version catalog accessor classes that Gradle generates at sync.

Now, for K2, the navigation could work even without accessors - see K2GradleTomlVersionCatalogReferenceProviderContributor. However, it's still required to sync a project with at least an empty version catalog to save its location in WSM. Having this location, it's possible to get the TOML file and find in its PSI the corresponding target element for navigation. Even if it was added or renamed but not synced yet.

GitOrigin-RevId: 040040f655bb906f20719ce0dc46077571ca1aab