[UI] IJPL-192439 Enable the background image for the main toolbar
It was initially removed because it wasn't supported for the entire
header. Now that this has changed, we need to enable it back
because it doesn't look good when the toolbar is in a separate
row, the header has the background image, but the toolbar does not.


(cherry picked from commit 4d674661d6fe7cdbc7a3b0fa50b98f5f16ca5fa6)

IJ-CR-180802

GitOrigin-RevId: f34b7e9199b69f9e8281a6ebc6adf802046bf0d3