PY-85358: Disable PythonSdkConfigurator for the Welcome project
Since there's an explicit configuration of SDK for the welcome project,
we don't need to use SDK autodetection logic. If we do, we end up in a
situation when we have two SDKs (and potentially a race when we choose
one).


(cherry picked from commit e6cf16d871ec000a6127018c156e53d032219d07)

IJ-MR-181358

GitOrigin-RevId: 4200c95af49ec4b311a62f446118c5bf68389f08