[terminal] IJPL-215631 Report additional data about popup completion in Terminal to FUS
Report terminal command context that will help to understand how effective is completion for any particular command.
Added our own `terminal_last_selected...` fields that should always be set when popup is closing. Because common completion fields report information about the selected element only if it was chosen explicitly or fully typed. It doesn't report any information about the last selected item if completion was canceled explicitly or on typing. But we need this information to better evaluate the relevance of shown suggestions.

Also, change `explicit` value parameter in `Lookup#hideLookup` calls to `false`. Because otherwise, it will be reported as `CANCELLED_EXPLICITLY` instead of `CANCELLED_ON_TYPING`, which doesn't sound correct. As our logic is closing the popup mostly in the cases when typings do not match the completion items.

GitOrigin-RevId: 1e7451367d367ac7b3206d71416fad28a44a4e9b