PY-84969 Make sure sdkAdded callback is called for remote interpreters
After changing dialogs for creating remote SDKs we ended up in a
situation when we've already persisted our SDK (see commonKt#setupSdk).
That leads to a situation when we're not calling our sdkCreated callback
in exitHandler of TargetEnvironmentWizard (see AddInterpreterActions).
There is a check for existing SDK (and it was already persisted), and
only if SDK doesn't exist we execute the callback. And if we don't
invoke that callback, we're also not getting to the jdkAdded call in
PythonSdkTableListener, which in the end leads to a missing SDK entry in
IDEA after creation (it only appears after reopening Project Structure).

The fix is to move the sdkCreated callback out of the condition, however
we don't really need the calls to addSdk and apply since we do addSdk
inside the callback anyway and it's not necessary to call apply since it
only affects the changelog of sdks for Apply button in Project
Structure.


(cherry picked from commit 56ec8af1889117503d19aeae9c95fe3f930de95d)

IJ-MR-181485

GitOrigin-RevId: f3433d61aa8e17b9e64b7a1293266e1317240450