Bundle the Natural Languages plugin
The Natural Languages plugin was formerly known as the Grazie plugin.
It was originally excluded in Change Ice27e8407 due to its dependency
on the Markdown plugin, but we bundle the Markdown plugin now
(and the dependency on Markdown is optional now too).

In IntelliJ 2025.3, some spellchecker functionality moved into the
Natural Languages plugin, and JetBrains recommended that we bundle it
in Studio.

Fixes: 444470096
Change-Id: Ibac988128a65a71b54826f9ad7d83da778d10259