[clion][debugger] CPP-47206 fix source kind selection
Before the fix for CPP-46877, XDebugSession.currentSourceKind was
used for most execution position updates.
The sourceKind based on XAlternativeSourceHandler.isAlternativeSourceKindPreferred
was used only in the positionReached() handling which is similar to
the sessionPaused() event.

After the fix for CPP-46877, we started to use the source kind based
on XAlternativeSourceHandler.isAlternativeSourceKindPreferred for
every position update.

Restoring the previous logic fixes the issue with switching to disasm.

GitOrigin-RevId: adae716559248c625bd88053ce041177560c8001