[java-tests] IDEA-381559 fix DataProvider not being set on TestApplicationManager
This fixes failing tests because of CodeInsightTestFixture#renameElementAtCaretUsingHandler throwing "No handler for this context".

BasePlatformTestCase#setUp calls (on line 41) BasePlatformTestCase#createMyFixture, which creates a CodeInsightTestFixtureImpl.
The myProjectFixture field of this CodeInsightTestFixtureImpl is set to an instance of LightIdeaTestFixtureImpl.

Then, on line 44, it calls CodeInsightTestFixtureImpl#setUp, and it ends up setting a TestDataProvider:

CodeInsightTestFixtureImpl#setUp
 \LightIdeaTestFixtureImpl#setUp
  \ ApplicationManager#setDataProvider
   \ HeadlessDataManager#setTestDataProviderDataProvider

But in JUnit 5, codeInsightFixture wasn't calling setTestDataProvider.
Because of this, TestDataProvider#getData was not being called in the process of creating DataContext.
As a result, DataContext had empty editor (CommonDataKeys.EDITOR), so it also had an empty PsiElement (CommonDataKeys.PSI_ELEMENT).

GitOrigin-RevId: 086e78812bc8d26b4d75c45bad5120a8ce730b73