[UI, plugins] IJPL-60552 Introduce JBScrollPane.verticalScrollBarNeedsSpace()
This is needed for cases when the client code needs
to make a decision based on whether the vertical
scrollbar changes the width of the viewport.

The particular case where it's needed right now
is the plugin description component,
where horizontal resizing causes vertical resizing
(proportional image scaling), sometimes
leading to an endless resize loop.

GitOrigin-RevId: f5d4861a1884afd67c2d9d16769e5f4a0f7011fa