[UI, plugins] IJPL-60552 Fix endless plugin description resize
Make the vertical scroll bar permanently visible
in configurations where it affects the scroll pane's width.
This prevents the endless loop when the image is shrunk
because the vertical scroll bar appears, which makes the
vertical scroll bar disappear (because proportional resizing),
which makes the image larger again and so on.

GitOrigin-RevId: c3cda0670e2f96a8df0895111bf3822a3cd13d01