IJPL-200859 Use JBZipFile instead of ZipFile to test the results of modification via JBZipFile
For some reason ZipFile cannot open a zip file which was generated with commons-compress library if Zip64Mode.Always mode is used ('Invalid CEN header' error is thrown if jdk 25 is used). Even if the file was not modified or opened with JBZipFile.
Library update didn't help. How we generate the zip also seems to be fine.

It was decided to use JBZipFile to open the file which was modified by JBZipFile

GitOrigin-RevId: 1887811ae35f880f4c80d596436504cc9a79b25b