[UI, islands] IJPL-217388 Fix the tool window stripe drop area highlight
Use the same padding that is used to paint the actual buttons.
This way it's always aligned with the buttons.

A side effect is that it now has the same size as the buttons
even with tool window names on, but that is likely a good thing.


(cherry picked from commit 658f1cca2cd2f56d2096cc0cdb6b6b3c5e84fc10)

IJ-CR-181870

GitOrigin-RevId: 94986e26f5b3c7c35e65369fa96038338c408d50