[multiverse] IJPL-217347 restart async commit if another FileViewProvider is created during commit run
Imagine the following situation:
There is a file with 2 code insight contexts.
And FileManager has a PsiFile with context A for the file.
Now, the file text gets updated and is being asynchronously committed.
At the same time, another activity spawns a read-action that requests a PsiFile of this file with context B.
The requested PsiFile must stay in sync with the document.

GitOrigin-RevId: 080dcd87964a5e3d4e3a5abdaf078fe360e96a12