Ignore missing platform test data
Some test fixtures such as JavaCodeInsightTestFixture expect to find a
mock JDK under intellij-community/java/mockJDK-1.7 (for example).
Studio unit tests have been passing with this test data missing for a
while, but IntelliJ 2025.3 now throws a hard error when any platform
test data is missing. We will ignore the hard error for now.

Another solution is to completely avoid JavaCodeInsightTestFixture
(and related test fixtures) in our tests, but this is a larger task
that would need to wait until after the IJ 2025.3 update has landed.
Or, if we end up integrating with the platform Bazel build, we could
try using the platform mock JDK directly.

This reapplies change I5931cd982 following merge conflicts with
IntelliJ 2026.1.

Bug: 458767289
Test: various Bazel tests with IntelliJ 2025.3
Change-Id: I365dacc7bf6697e204773044250c5faf8bb4f386