[UI, threading] IJPL-189188 Add @RequiresReadLock to some PsiCopyPasteManager methods
Sometimes they're called directly by AWT bypassing locks.
And these days locks aren't guaranteed on the EDT anyway.
So let's annotate and wrap into RA/WIRA as appropriate.
Mostly it's just a RA, but in the case of the paste action
it's quite possible that it'll want to write something as well,
so let's make it a WIRA there.

GitOrigin-RevId: f9d05b1ae4691f39d9f2740173e90ee5ea8b3768