Add action to toggle Show Layout Bounds in Compose UI
This adds an action to the ui-inspector module that toggles the
showLayoutBounds value for all currently attached ComposePanels.
It also registers an AWT hierarchy listener, if the toggle is
on, to capture when a new ComposePanel is added to the hierarchy
and make sure the setting is respected in cases where panels are
detached and then re-attached (the setting is lost when the
panel is detached as it disposes a lot of internal state).

When toggled, the action also tries to cause a recomposition in
all the panels so that the bounds do start showing right away,
but it doesn't seem to always work/cause a full redraw. Any
interaction or change in bounds on the panels will cause the
panels to redraw, and that is mostly good enough for now.

Given the approach is reflection-heavy, it is prone to breaking
if there are changes in CMP internals; when we have a better,
official API for it, we'll be able to have a more robust impl.