DO-1385 [statistics]: fix issue where setting custom path for metadata in statistics tool window had no effect.
This problem was introduced by migrating some of the metadata logic into the reporting SDK. The functionality used to be a special case of the regular metadata storage logic.

We fix this by making it part of the test validation rule storage. This makes more sense because the feature is solely for testing purposes.

GitOrigin-RevId: 96c03e75bc6549c361e2669bec2e48147310e326