IDEA-382020 [java-inspections]: 'Split declaration and initializer' produces exception when injection is used
Analysis: `CodeStyleManager::reformat` inside ModCommand triggers reformatting of the injected fragment as well (IJPL-35687), but due to implementation deficiency, the injected fragment always has the event system enabled, even for a non-physical host file. Fixing this in the platform looks hard, but we can avoid reformatting the injection inside the intention.
Implementation-wise, I used the CodeBlockSurrounder which does the similar thing but does not have the reformatting problem. I also improved CodeBlockSurrounder to support corner cases and added more tests.

GitOrigin-RevId: c7467aa5771489c95698dc01b487059d87d655c3