IJPL-218391: change the API status of FloatingToolbarProvider.onHiddenByEsc
This will allow the plugins to implement FloatingToolbarProvider.

Without this change, this is currently forbidden by the API Verifier because Kotlin will emit a synthetic method when implementing an interface with a default method, which will be (incorrectly) seen as usage of the internal method in question.

See also: MP-5395.

#IJPL-218391 Fixed

GitOrigin-RevId: 2e8f102115ba31146701412f26233362ea418059