Fix included GradleBuild conversion to DefaultGradleLightBuild
With IntelliJ 2025.3 more in particular ce7ec3afd6a17c6016bf commit,
some refactor has done for composite builds as part of IDEA-370426
to add GradleLightProject.getIdentityPath. However, the GradleBuild
class which doesn't implement hashCode() is defined as key map
for gradleBuildsToConverted in DefaultGradleLightBuild.

This fact is causing NullPointerException with Gradle 6.X since
a different instance is created for same build path, failing
to map included GradleBuild into DefaultGradleLightBuild,
which was reported on IDEA-382120.

This change corresponds to a pending pull request at
https://github.com/JetBrains/intellij-community/pull/3304.

Bug: 459390354
Bug: 460707123
Test: Existing
Change-Id: I553a180870f837d7ae3aa8b61436eb5e1af5e188