PY-85452 Persist sdk for remote interpreters and upload proper paths
Firstly, we have JDK table where we store SDKs, but also we have our own
ProjectSdksModel for interpreters. If we won't persist SDK in our model,
but only in JDK table, we'll technically have SDK, but won't set it
properly in different places (eventually we'll sync those models
though). So it's important to persist SDK using ProjectSdksModel.

Secondly, for SSH interpreters we used helpers mappings for all uploaded
files. The proper way is to detect whether the file is a helper and
apply appropriate mapping.

These changes address both these issues.

GitOrigin-RevId: d582fc714326324c0580f11b13fd1f843d8ebefa