refactoring [Code Insight]: Refactor PY-85639 the test framework implementation
The main change is introducing the `PyCodeInsightTestApplication`, unfortunately, we cannot reuse the old `PyDefaultTestApplication` because it does not isolate the test data properly, specifically, pre-loads all the test data into the temporary dir.
The new approach loads only the data related to the particular test (file/sub dir if it's a multi-file test) or nothing if it's an in-place test

Refactor JUnit 5 utilities for Python inspections and annotations

- Moved utility classes to appropriate packages under `impl` or `annotations`.
- Introduced new annotations: `InspectionTest`, `WithLanguageLevel`, `InjectCodeInsightTestFixture`, and `PyCodeInsightTestApplication`.
- Created dedicated extensions: `PyTestDataExtension`, `PyWithLanguageLevelExtension`, and `PyInspectionTestJUnit5Extension`.
- Enhanced modularity and organization of multi-file and code insight-related logic.
- Updated references in `PyMissingConstructorTest` to reflect changes.

Restructure `PyTypeCheckerInspection` test data and related JUnit 5 test file. Add explanatory comments

Space-RevId: 0fbc7623a89f12d01520fb7f326a9e0a512c8bf9

GitOrigin-RevId: bfd2990a5b13fa83c080a208ba336f2821c1f3ac