IJPL-64483 Indent cursor to the parent of sequence when pressing Enter twice
The behavior for Enter in the middle item has changed—now it also indents the cursor to the parent of sequence. This was done as it seems that there is no single meaningful behavior in the case when user presses Enter twice in the middle of sequence, so it's better to keep things simpler and consistent.

Also fixes AMPER-57.

GitOrigin-RevId: cd6c77d8406afd30957441c8d6b44a7c8292dcf2