IJPL-177151 Incorrect highlighting when commenting
When the map eviction listener for the entire file is run, eviction listeners for all maps beneath VISITED_PSI_ELEMENTS that are just collected too (e.g. ToolHighlights.elementHighlights) sometimes are not run. If the daemon restarted in this precise moment, there's a risk for dangling range highlighters. It means the "comment" action will remove the highlighters registered in data, but the dangling ones will be still shown, creating this "highlighted parts inside comment" effect. To fix, we'll gather all highlighters from the document markup and delete them manually, when the host psi file is gc-ed.

GitOrigin-RevId: 8bf207e1cedd369ff7fdc4c366003b76da097460