[UI] IJPL-218177 Fix the non-fullscreen frame border on macOS
The code in setStoredFullScreen always set the border to null,
regardless of the full screen state. Apparently after moving
this invocation to a later stage this now causes issues
in non-full screen. Previously it was likely overridden
by some code that is now executed earlier than this.

Fix by only removing the border in full screen.

GitOrigin-RevId: 8b3a213454479d5a26b398a04a0b4141cac0900a