[UI, PV] IJPL-197345 PV performance calculation
Because update requests can run in parallel,
summing up their times would yield ridiculous
results like 5 minutes per minute.
So we track individual update causes from the start
of the first request to the end of the last request
related to that cause. This gives us a consistent
"How much time did the PV spend serving these requests?"
metric.

We also do not report incomplete requests every
minute, because that would be confusing.
Imagine a faulty request that runs 2.5 minutes.
It would be reported three times as having spent
1, 2 and 2.5 minutes. Instead, we report it only once
it ends. To give us some idea about long-running
requests we also keep track of the total number of
active requests.

For now reporting is only done into the log.
We'll add FUS and/or telemetry later.

GitOrigin-RevId: dd345f3ae65be57b8f03259f9d22237a4a2fc619