[gradle][wsm] IDEA-381806 GradleProjectEntity symbolic ID: use identityPath instead of url
In the Kotlin project, there is a case when the same directory of the same build could be used by two different Gradle projects. IDEA throws an error in this case because a symbolic ID for GradleProjectEntity is duplicated. To make it unique, I replaced a directory URL with an identity path.

Tests:
* GradleContentRootSyncContributorTest

GitOrigin-RevId: 388410ceff9b7ef29a89c3f1fb3e0277c80d9340