[JEWEL-1142] Ignore Cursor summaries in release notes script
Since the Cursor bot has recently been added to the intellij-community
repo, this has broken the release notes script. The bot appends a
"summary" to each PR description, making so that the release notes are
not the last part in the PR description, and breaking the script logic.

Since the Cursor slop/summary is contained in `<!-- CURSOR_SUMMARY -->`
comments, and it's always appended to the human-provided PR description,
we can safely ignore everything from that comment onwards when we
extract the release notes from PR bodies.

closes https://github.com/JetBrains/intellij-community/pull/3308

GitOrigin-RevId: d5e93bcdc7ef4c23ced437dd4d86ac9cad15013b