[testFramework] IJPL-216806: Let slow runnables terminate if a leak is detected
Sometimes we can have a rogue computation that is doing something while we check for leaks. We want to wait for reasonably fast computations, so we give them a chance to finish when we first detect a leak. If the leaking object remains, then either we found a true leak, or a computation does not cancel itself fast enough.

GitOrigin-RevId: cb811cd62644a8efdcb6ef6addb8f0096b8f5074