[UI, Find Usages] IJPL-162332 Cache canNavigate for usages
Let's try to cache the values to implement canNavigate
so it can be safely invoked on the EDT.
This should fix the slow ops in next / prev occurrence.
There's a risk of performance regressions, though,
because now this value will be precomputed for all nodes,
and there can easily be thousands of them.

GitOrigin-RevId: 068f3f99c600024aea1aab9b5eba7e140ca951e3