testFramework [projectStructure]: Fix subDirFixture dependency
The subDirFixture created the subdirectory by the dirFixture, which was not dependent on the superdirectory fixture. This caused the deletion of the super and sub directories to be done concurrently, which caused problems on Windows.
This refactor creates and cleans up the directories manually directly in the fixture, so that the dependency is set up correctly.

GitOrigin-RevId: 7fbde6449b0a51ca2bbaa76fabcfe4c62c07923f