[terminal] IJPL-215394 Update command start offset of block when prompt text changes
If text changes inside the prompt after command start offset was set, we need to update it, otherwise features that rely on it (like typeahead) might work incorrectly.
Unfortunately, can't predict how to update offsets when a change happens both in the prompt and command text. So, in this case, offsets will be wrong.

GitOrigin-RevId: bd9ae3e04acd8036a5ea8f9d194d1fb520132146