[terminal] IJPL-218114 Fix removing prompt by type-ahead after invoking Clear terminal action (Cmd+K)
Cmd+K clear action is quite fragile because it is implemented in the terminal emulator side. It is just clearing the buffer without letting the shell process know about it.
And unfortunately, we can't properly handle it in the `TerminalBlocksModel` - the text change event is just a full replace that breaks the current block offsets, so we are no more aware of what are the prompt and command positions there. See no way to fix it.
But, we can add protection for this case to the type-ahead logic - to disable type-ahead in such block, and avoid false-positive typing.


(cherry picked from commit f544069f83ba0c4e1db4b78a9fdacbece19147e7)

IJ-CR-183173

GitOrigin-RevId: 91f431c38b866ad9c99927816b3f5fa705821744