testFramework [projectStructure]: Fix multiverse project path fixture
The MultiverseFixtureInitializer created a `dirFixture` based on the project base path. The project fixture itself created a `tempPathFixture` for the project directory. This caused the project directory to be deleted twice without a relation.
This commit manually creates the `tempPathFixture` and passes it both to the project fixture and to the directory initialization logic.
Modified to omit logging for 253 cherrypick.

GitOrigin-RevId: a3d613490e0fdfcb32f07aa47c148b695bc49e54