refactor [lsp]: abstract over i/o to remove ktor dependency
We want to remove ktor dependency from the lsp.protocol module as it
conflicts with other uses.

The module has a single usage of ktor, its i/o utils types:
ByteReadChannel and ByteWriteChannel.

These are now replaced by requivalent interfaces. In this patch, ktor
dependency is still present and implements the aforementioned interfaces
in this same module, but the goal is to move the implementation into
module consumers.

GitOrigin-RevId: 62681cd01c54b00cab7e17682fe6858db01d4186