IJPL-169109 build scripts: search for forbidden paths only in package names
This is classes and packages names that are scrambled, not JAR names, so it should be ok to have e.g. 'license' in the JAR name. And we'll get such a JAR name after extracting licensing-related code to a separate content module put to a separate JAR as part of IJPL-169109.

GitOrigin-RevId: a9d9cd2e7ba7441f5e302bc05bd2da7997bdcf3b