[IJPL-219340] Fix UnionScope.tryUniting for case when 1 scope is left
When `leftScopes` only has one scope remaining, the code calls back into
`tryUniting`, which requires at least two scopes. Since we know that the
single remaining scope can't be united with the scope that it just tried
to unite with, it's okay to just return that united scope with the
remaining one.

Pending PR: https://github.com/JetBrains/intellij-community/pull/3315

Bug: 461785890
Test: ProGuardR8ClassReferencesIntegrationTest
Change-Id: I2630475e56c6bd6c3ee3f870658f3e54c4a38945