[IJPL-219340] Fix UnionScope.tryUniting for case when 1 scope is left
https://youtrack.jetbrains.com/issue/IJPL-219340

When `leftScopes` only has one scope remaining, the code calls back into
`tryUniting`, which requires at least two scopes. Since we know that the
single remaining scope can't be united with the scope that it just tried
to unite with, it's okay to just return that united scope with the
remaining one.

GitOrigin-RevId: 16d15ba991e583aa903da7e3840dfc6dd12c0e7f