[eel] IJPL-218598: Make Gradle obey `JAVA_HOME` even when set locally and pass it for tests.
Currently, Gradle uses `JAVA_HOME` to detect a JDK. But it works for local setups only, so Docker/WSL Gradle test failed for 2 reasons:
1. We didn't provide JDK path as `JAVA_HOME`
2. Gradle didn't use `JAVA_HOME` for anything but local eel (see `org.jetbrains.plugins.gradle.util.GradleJvmResolutionUtil.canUseJavaHomeJdk`)

The right solution could be the following:
1. Fix Gradle to ask the Eel for `JAVA_HOME`
2. Patch user environment to provide correct `JAVA_HOME`

The latter one can't be fixed anyway, as we do not control the environment (not WSL at least).
The former will be fixed soon.

After the extensive discussion with A. Glukhov, I've come to the following solution:
1. We set `JAVA_HOME` as a host env (i.e `\\wsl$\` or `/docker/..` for remote eels).
2. We make Eel obey it (but only if points to the Eel of course)


Merge-request: IJ-MR-182515
Merged-by: Ilya Kazakevich <ilya.kazakevich@jetbrains.com>

GitOrigin-RevId: 8971baf7a6ac831cd6d46acacc6725da1de451f6