[terminal] IJPL-213457 Provide initial query text for Search Everywhere and Find In Files
Previously it was working because we exposed the editor into the data context, and features like SE and Find In Files were able to get the text selection from it and use it as an initial query. Now we do not expose the editor, so we need to manually provide the selection text to the data context.


(cherry picked from commit ff3581c023fb2e2d6e1ea25dc6aaa144648db9f6)

IJ-CR-183306

GitOrigin-RevId: 5e72fcf63811300cc34e91ec5d12f9880b2b95ae