[terminal] IJPL-219116 Fix duplicate hyperlinks in the terminal
When a new task is created because the model has been
updated, it's put into the runner state flow, which causes
the collector to trigger. This sometimes causes
the collector to run the previous task again.

Add a test and fix using distinctUntilChanged
to prevent processing of the same task.

GitOrigin-RevId: 28c51d65c596021076d206a551bbcb2482f46fd7