[terminal] IJPL-219116 Fix duplicate hyperlinks in the terminal
When a new task is created because the model has been
updated, it's put into the runner state flow, which causes
the collector to trigger. This sometimes causes
the collector to run the previous task again.

Add a test and fix using distinctUntilChanged
to prevent processing of the same task.


(cherry picked from commit 28c51d65c596021076d206a551bbcb2482f46fd7)

IJ-CR-183360

GitOrigin-RevId: 289eaa30dfcc1e2d28eea936cbf33bab3754851e