[rd debugger] Introduced XDebugSessionProxy API to get execution stacks without suspendContext.
Note: to provide a short-living scope for execution stacks produced without suspendContext, the child coroutineScope of the debug session scope was created, which is canceled on resume/pause/stop debug session events. Execution stacks are hanged to this child scope and are cleaned up as soon as the scope is disposed.

IJPL-203827

GitOrigin-RevId: 11e958bf6e5af9fc5be9deed46f77edf6295280c