[kotlin] KTIJ-34539 Start to suggest auto-imports of type aliased inner classes when appropriate
Implement `TypeAliasedInnerClassImportCandidatesProvider` to provide relevant candidates
and check for the applicability with know receivers.

Also, adjust the implementation of the `ImportQuickFix`
to avoid inserting full qualifiers for such references,
see the KDoc on the `isTypeAliasedInnerClassUsedInCall`
for the reasoning

Known limitation: type parameters substituted at the side of the typealiases
are not properly checked, see the KDoc on the `canBeCalledAsInnerClassConstructor`
function for this.

The issue mentions nested type aliases; they are naturally supported
along with all other cases; no extra logic is required.

This implementation works only in K2 Mode.

^KTIJ-34539 Fixed

GitOrigin-RevId: a7d2baf819123e9d31df72c3043456c518a9f520