[completion] IJPL-207762 add support for prefix matchers
PrefixMatchers can contain pretty complex logic which might be missing on the Frontend,
thus there must be a way to retrieve its results from the Backend.

To support that, FrontendCompletionPrefixChangeListener is introduced.
It sends prefix updates to the Backend, and the Backend responds with the matching update.
This also allows resorting items according to their weights.

GitOrigin-RevId: 53e8b8d48ee0f08ab4edb3b1c1af03791fcb4333