[terminal] IJPL-198669 Fix broken text in terminal editor tab when switching between tabs
It was an exploding head mystery.
Actually, the text was looking like 1 char on a line because of soft wraps implementation based on the `CharacterGridImpl` we use in the terminal editor.
It happened because `CharacterGridImpl.getSize` was zero for the terminal editor when it was hidden (other editor tab is opened). So, soft wraps were placed after every character for all new terminal output.
The editor had zero size because of `JBTabsImpl.applyResetComponents` that sets the bounds to zero for the previously selected editor content component (it is removed from the UI hierarchy later).
So, the fix is to not reset bounds to zero for `TabInfo.component` components. It looks useless and causes the problem in the terminal. Every tab content component will have `LAYOUT_DONE` set once during first layout and never reset.


(cherry picked from commit ad119904f743e4583d5642499104158fd9e8a908)

IJ-CR-183641

GitOrigin-RevId: 3052742740c04371cb71d5fcfe1cf6653d3cce45